+++ shoop/core/models/_suppliers.py
@@ -65,6 +65,21 @@ class Supplier(ModuleInterface, ShoopModel):
         """
         return self.module.get_stock_status(product_id)
 
+    def get_suppliable_products(self, shop, customer):
+        """
+        :param shop: Shop to check for suppliability
+        :type shop: shoop.core.models.Shop
+        :param customer: Customer contact to check for suppliability
+        :type customer: shoop.core.models.Contact
+        :rtype: list[int]
+        """
+        return [
+            shop_product.pk
+            for shop_product
+            in self.shop_products.filter(shop=shop)
+            if shop_product.is_orderable(self, customer, shop_product.minimum_purchase_quantity)
+        ]
+
     def adjust_stock(self, product_id, delta, created_by=None):
         return self.module.adjust_stock(product_id, delta, created_by=created_by)